refactor: move positive double from model to utils
