+++ src/main/scala/io/github/srs/utils/types/PositiveDouble.scala
@@ -1,4 +1,4 @@
-package io.github.srs.model
+package io.github.srs.utils.types
 
 import io.github.srs.model.validation.Validation
 
@@ -22,7 +22,7 @@ object PositiveDouble:
      *   [[Right]] if the instance is valid, or a [[Left]] with an error message if invalid.
      */
     def validate: Validation[PositiveDouble] =
-      import io.github.srs.model.validation.Validation.{ notInfinite, notNaN, positive }
+      import io.github.srs.model.validation.Validation.{notInfinite, notNaN, positive}
       for
         value <- notNaN("value", pd)
         value <- notInfinite("value", value)
